app/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "streamlit_app"
version = "1.0.0"
description = ""
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "streamlit==1.31.0",
    "pytube==15.0.0",
    "requests==2.27.1",
    "deepgram-sdk==3.1.2",
    "matplotlib==3.8.2"
]

[project.scripts]
streamlit_app = "main:main"